Move errors into their appropriate module
